Task-4

+++ database/migrations/2024_05_07_181503_create_students_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('students', function (Blueprint $table) {
             $table->id();
             $table->string('studentName', 100);
-            $table->integer('age');
+            $table->tinyInteger('age');
             $table->timestamps();
         });
     }
